Fixed misc UI bugs, and hopefully also the strange exception I encountered earlier.
The latter may be a bug in the JRE:
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=7075600
I have sidestepped the problem by forcing java to use the old mergesort instead of the newer 'timsort' algorithm.
Mergesort is often slower, but this isn't really noticable on the client.

I have added it in MedSavantClient.Main with:
System.setProperty("java.util.Arrays.useLegacyMergeSort", "true");

This could also be set via a command line parameter, if that's preferrable.
-Djava.util.Arrays.useLegacyMergeSort=true

Since this change, I haven't received the exception.