Temporary mechanism to restart MedSavant
This restarts via an exec() call. This needs to be tested on all
platforms to ensure that it works properly. It should be tested
specifically on OS-specific launchers (e.g. does the instance started
by exec have a nice icon?). In the future, we should remove static
variables or reset them appropriately.